iTPAEngine/__init__.py
import pytesseract
from paddleocr import PaddleOCR
from pdf2image import convert_from_path
from PIL import Image
import tempfile
import os
import re
import csv
import base64
import os
import json
from openai import OpenAI

# Exposure of main classes for the package
from .iTPAEngine import DamageAnalyzer, DocumentOCRProcessor

# You can also define what is exported when someone uses 'from package import *'
__all__ = ["DamageAnalyzer", "DocumentOCRProcessor"]
