requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
bce-python-sdk==0.9.56
bitsandbytes==0.49.0
black==25.12.0
bleach==6.3.0
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
contourpy==1.3.3
cut-cross-entropy==25.1.1
cycler==0.12.1
datasets==4.3.0
diffusers==0.36.0
dill==0.4.0
docstring_parser==0.17.0
filelock==3.20.1
filetype==1.2.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.9.0
future==1.0.0
h11==0.16.0
hf-xet==1.2.0
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.7
ImageHash==4.3.2
imagesize==1.4.1
importlib_metadata==8.7.1
Jinja2==3.1.6
kaggle==1.8.3
kagglesdk==0.1.14
kiwisolver==1.4.9
librt==0.7.5
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
modelscope==1.33.0
more-itertools==10.8.0
mpmath==1.3.0
msgspec==0.20.0
multidict==6.7.0
multiprocess==0.70.16
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.6.1
numba==0.63.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai-whisper==20250625
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==25.0
paddleocr==3.3.2
paddlepaddle==3.2.2
paddlex==3.3.12
pandas==2.3.3
pathspec==0.12.1
pdf2image==1.17.0
peft==0.18.0
pi_heif==1.1.1
pillow==12.0.0
pillow-avif-plugin==1.5.2
platformdirs==4.5.1
polars==1.36.1
polars-runtime-32==1.36.1
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
pypdfium2==5.2.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytokens==0.3.0
pytz==2025.2
PyWavelets==1.9.0
PyYAML==6.0.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
roboflow==1.2.11
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scipy==1.16.3
sentencepiece==0.2.1
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
text-unidecode==1.3
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
torchao==0.15.0
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
trl==0.24.0
typeguard==4.4.4
typer-slim==0.20.1
types-requests==2.32.4.20250913
types-tqdm==4.67.0.20250809
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==1.0.3
tzdata==2025.3
ujson==5.11.0
ultralytics==8.3.241
ultralytics-thop==2.0.18
unsloth==2025.12.9
unsloth_zoo==2025.12.7
urllib3==2.6.2
wcwidth==0.2.14
webencodings==0.5.1
wheel==0.45.1
xformers==0.0.33.post2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
